requirements.txt
asyncpg~=0.27.0
pydantic~=1.10.2
uvicorn~=0.20.0
fastapi~=0.87.0
python-dotenv~=0.21.0
python-multipart~=0.0.5
pytest~=7.2.0
pytest-asyncio~=0.20.2
asgiref~=3.5.2
python-jose~=3.3.0
jinja2~=3.1.2
bcrypt~=4.0.1